商品マスタ機能実装完了

+++ Project/src/ec/UserItemDelete.java
@@ -1,7 +1,6 @@
 package ec;
 
 import java.io.IOException;
-import java.util.ArrayList;
 
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
@@ -10,7 +9,6 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import javax.servlet.http.HttpSession;
 
-import beans.ItemDataBeans;
 import dao.ItemGetListDAO;
 
 /**
@@ -19,25 +17,26 @@ import dao.ItemGetListDAO;
 @WebServlet("/UserItemDelete")
 public class UserItemDelete extends HttpServlet {
 	private static final long serialVersionUID = 1L;
+
+
 	protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
 		HttpSession session = request.getSession();
 
 		try {
 			int userId = (int) session.getAttribute("userId");
 			String[] deleteItemIdList = request.getParameterValues("delete_item_id_list");
+			 String listActionMessage = "商品が選択されていません";
 			//ほしい物リストから削除する
+			if(deleteItemIdList!=null) {
 			for (String deleteItemId : deleteItemIdList) {
 				System.out.println(deleteItemId);
 			    boolean flag = ItemGetListDAO.getInstance().deleteUserItemByItemId(userId,Integer.parseInt(deleteItemId));
-			}
+			    listActionMessage = "ほしい物リストから商品を削除しました";
+			}}
 
 
-			String listActionMessage = "削除しました";
-			//ほしい物リストの取得
-			ArrayList<ItemDataBeans> userItemList = ItemGetListDAO.getInstance().getUserItemList(userId);
-			request.setAttribute("listActionMessage",listActionMessage);
-			request.setAttribute("userItemList", userItemList);
-			request.getRequestDispatcher(EcHelper.ITEM_GET_LIST_PAGE).forward(request, response);
+			session.setAttribute("listActionMessage",listActionMessage);
+			response.sendRedirect("ItemGetList");
 
 		} catch (Exception e) {
 			e.printStackTrace();
